pyproject.toml
[project]
name = "near-metadata-chatbot"
version = "0.1.0"
description = "Streamlit-based Excel metadata chatbot with semantic RAG."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.30",
    "pandas>=2.0",
    "python-dotenv>=1.0",
    "openpyxl>=3.1.2",
    "langchain-core>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-text-splitters>=0.2.0",
    "langchain-huggingface>=0.0.3",
    "langchain-groq>=0.1.0",
    "langchain-chroma>=0.1.0",
    "sentence-transformers>=2.2.2",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["semantic_rag_chatbot"]